Add Wala plugin consuming Kafka topic

+++ analyzer/javacg-wala/src/main/java/eu/fasten/analyzer/javacgwala/data/callgraph/PartialCallGraph.java
@@ -72,7 +72,7 @@ public class PartialCallGraph {
      *
      * @return FASTEN call graph
      */
-    public RevisionCallGraph toRevisionCallGraph(int date) {
+    public RevisionCallGraph toRevisionCallGraph(long date) {
 
         List<List<RevisionCallGraph.Dependency>> depArray = new ArrayList<>(coordinates.size());
